抽离demo到独立文件，同时去除mockjs，调用yapi
