+++ package.json
@@ -48,7 +48,6 @@
     "@storybook/addon-notes": "^3.4.8",
     "@storybook/addons": "^3.4.8",
     "@vue/test-utils": "^1.0.0-beta.16",
-    "mockjs": "^1.0.1-beta3",
     "babel-jest": "^23.0.1",
     "babel-core": "^6.26.3",
     "babel-preset-env": "^1.7.0",
